Ensure post-rendering bookkeeping is always performed (#532)
This is primarily a fix for a bug in `ExtAudioFile` where
`ExtAudioFileRead` returns zero frames despite not reaching EOF,
(documentation to the contrary notwithstanding).

It also addresses a situation where decoding is complete but zero frames
are available in the ring buffer, so the rendering complete notification
would never be sent.