Back off direct use of `_playerNode->_impl`
Since `_playerNode` may be swapped out by
`-configureProcessingGraphForFormat:forceUpdate:`, there is a small
chance that it could be nil if one of the other methods on
`SFBAudioPlayer` are called between the time `_playerNode` is set to nil
and reassigned. Sticking to Objective-C takes advantage of the fact that
messaging `nil` isn't an error.