Back off direct use of `_playerNode->_impl` (#544)
Since `_playerNode` may be swapped out in
`-configureProcessingGraphForFormat:forceUpdate:`, there is a small
chance that it could be `nil` when used in one of the other
`SFBAudioPlayer` methods if that method is called between the time
`_playerNode` is set to `nil` and reassigned. Sticking to Objective-C
takes advantage of the fact that messaging `nil` isn't an error.