Reconfigure the audio processing graph during enqueue if needed (#600)
As described in #598 if a decoder is enqueued on a player when the
player's internal decoder queue is empty and the decoder's processing
format is not supported by the player node, the decoder is incorrectly
pushed to the internal queue when the player node does not have a
current decoder.

This PR instead reconfigures the processing graph with the decoder's
processing format and then enqueues the decoder on the player node when
the player node does not have a current decoder.