Remove audio engine isolation queue (#618)
No documentation for `AVAudioEngine` that I can find states that the
object may only be used from a single thread/queue. On the other hand,
no documentation that I can find states that it is safe to use the
object from multiple threads/queues either.

`AVAudioEngine` appears to use mutexen internally to protect graph
modifications, so this PR removes the extra logic and overhead
associated with the audio engine isolation dispatch queue.